+++ libft_mine/source/custom/ft_putchar_len.c
@@ -1,19 +1,19 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_putchar_fd_len.c                                :+:      :+:    :+:   */
+/*   ft_putchar_len.c                                   :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: donghyu2 <donghyu2@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/23 15:11:44 by donghyu2          #+#    #+#             */
-/*   Updated: 2022/12/02 01:36:14 by donghyu2         ###   ########.fr       */
+/*   Updated: 2022/12/11 18:41:32 by donghyu2         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <unistd.h>
 #include "libft.h"
 
-size_t	ft_putchar_len(char c)
+int	ft_putchar_len(char c)
 {
 	return (write(1, &c, 1));
 }
